moved SelectedItemId to model module
