+++ feature/timeline/src/main/java/com/freshdigitable/udonroad2/timeline/TweetListItemClickListener.kt
@@ -1,6 +1,7 @@
 package com.freshdigitable.udonroad2.timeline
 
 import androidx.databinding.ObservableField
+import com.freshdigitable.udonroad2.model.SelectedItemId
 import com.freshdigitable.udonroad2.model.Tweet
 import com.freshdigitable.udonroad2.model.TweetListItem
 import com.freshdigitable.udonroad2.model.TweetingUser
@@ -17,14 +18,4 @@ interface TweetListItemClickListener : ListItemClickListener<TweetListItem> {
 
 interface TweetListEventListener {
     val selectedItemId: ObservableField<SelectedItemId?>
-}
-
-data class SelectedItemId(
-    val originalId: Long,
-    val quoteId: Long? = null
-) {
-    @JvmOverloads
-    fun equalsTo(originalId: Long, quoteId: Long? = null): Boolean {
-        return this.originalId == originalId && this.quoteId == quoteId
-    }
 }
