+++ model/src/main/java/com/freshdigitable/udonroad2/model/SelectedItemId.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2020. Matsuda, Akihit (akihito104)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.freshdigitable.udonroad2.model
+
+data class SelectedItemId(
+    val originalId: Long,
+    val quoteId: Long? = null
+) {
+    @JvmOverloads
+    fun equalsTo(originalId: Long, quoteId: Long? = null): Boolean {
+        return this.originalId == originalId && this.quoteId == quoteId
+    }
+}
